+++ src/bot/constants/about_fund_text.py
@@ -1,0 +1,126 @@
+"""Файл с константами для блока "О Фонде"
+"""
+from bot.constants.links import (
+    ANNUAL_REPORTS_LINK,
+    PROCESSES_LINK
+)
+
+
+ABOUT_FUND_HISTORY = {
+    'msg_1': 'Для начала немного предыстории 👀',
+    'msg_2': (
+        'Весной 2013 года во дворе собора St. Andrews прошла '
+        'благотворительная распродажа в поддержку Центра равных '
+        'возможностей «Вверх». Несколько девушек собрали свою '
+        'одежду и продали её, а вырученные средства направили в «Вверх». '
+        'С этой акции началась наша история – спустя год её инициатор '
+        ' Дарья Алексеева открыла постоянно работающий '
+        'магазин Charity Shop, а еще спустя год в ноябре 2015 года был '
+        'зарегистрирован фонд ВТОРОЕ ДЫХАНИЕ.'
+    ),
+    'msg_3': (
+        'Сегодня фонд Второе дыхание – это крупнейшая некоммерческая '
+        'организация в России, занимающаяся сбором, сортировкой, '
+        'перераспределением и переработкой одежды на любой стадии ее жизни.'
+    )
+}
+
+FUND_MISSION = {
+    'msg_1': 'Здорово, что тебе интересно! Начну рассказ с важных цифр.',
+    'msg_2': (
+        'Ежегодно в России выбрасывают от 60 до 70 миллионов тонн твердых '
+        'бытовых отходов. Одежда и текстильные отходы составляют от 3 до 7% '
+        'на мусорных полигонах – то есть по меньшей мере 2 миллиона тонн '
+        'одежды ежегодно оказывается на свалках. Одежда разлагается до 200 '
+        'лет, и в процессе выделяет метан, загрязняет почву и грунтовые воды '
+        'и увеличивает объем мусорных полигонов. 😔'
+    ),
+    'msg_3': (
+        'Фонд ВТОРОЕ ДЫХАНИЕ старается предотвратить попадание одежды на '
+        'полигоны и создаёт инфраструктуру приема ненужных вещей, которые '
+        'при грамотном и эффективном использовании могут стать ценным '
+        'ресурсом для решения социальных и экологических проблем.'
+    ),
+    'msg_4': (
+        'Мы стремимся к развитию системы повторного использования и '
+        'переработки текстиля с 1% (столько собиралось в стране в 2022 '
+        'году от общей массы выброшенного) до 50%.'
+    ),
+    'msg_5': (
+        'Наша миссия – ОСОЗНАННОЕ ПОТРЕБЛЕНИЕ КАК ОБРАЗ ЖИЗНИ.'
+    ),
+    'msg_6': (
+        'Хочешь узнать, куда направляются вещи, которые сдают в Фонд?'
+    )
+}
+
+THINGS_PATH = {
+    'msg_1': 'Мы дарим старым вещам новую жизнь несколькими способами👇',
+    'msg_2': (
+        'Вещи, собранные Фондом и пригодные к повторному использованию, '
+        'передаются на благотворительность – нуждающимся семьям и людям, '
+        'попавшим в сложную жизненную ситуацию. Мы оказываем '
+        'благотворительную помощь в сотрудничестве с организациями, '
+        'которые работают с людьми, нуждающимися в поддержке, а также фондами '
+        'по защите животных и социальным учреждениями.'
+    ),
+    'msg_3': (
+        'Часть вещей продается в магазинах ВТОРОЕ ДЫХАНИЕ (Charity Shop) '
+        'в Москве, Ярославле, Костроме, Ростове Великом. Магазины нужны '
+        'для обеспечения собственных нужд фонда, выплаты заработной платы '
+        'сотрудникам и аренды помещений.'
+    ),
+    'msg_4': (
+        'Одежда, непригодная для дальнейшей носки, '
+        'отправляется на переработку.'
+    ),
+    'msg_5': (
+        'Показать, как выглядит анатомия процессов в Фонде, '
+        'чтобы было нагляднее?'
+    )
+}
+
+PROCESS_ANATOMY = (
+    'Иногда лучше 1 раз увидеть, чем 100 раз услышать, да?😄'
+    'Посмотри, '
+    f'*[как устроен процесс работы с одеждой]({PROCESSES_LINK}) '
+    'в Фонде\.\n'
+    'Рассказать какие еще проекты реализует наш Фонд?*'
+)
+
+FUND_PROJECTS = {
+    'msg_1': (
+        'Наш Фонд работает по принципу стартапа: мы тестируем'
+        ' гипотезы и принимаем те, которые сработали, а от тех, которые не'
+        ' дали нужный результат, отказываемся. Сейчас у Фонда есть несколько'
+        ' действующих проектов, рассказываю! 🤩'
+    ),
+    'msg_2': (
+        'В Фонде реализуется программа поддержки социальных предпринимателей,'
+        ' которая включает в себя обучающие мероприятия, консультации, обмен'
+        ' опытом и нетворкинг для социальных предпринимателей и НКО в '
+        ' регионах,которые работают в сфере сбора и перераспределения '
+        'текстиля.'
+
+    ),
+    'msg_3': (
+        'В 2021 году фонд ВТОРОЕ ДЫХАНИЕ открыл новое направление своей работы'
+        ' — общественную мастерскую в Москве. Это место, где любой желающий'
+        ' может починить сломавшиеся или устаревшие вещи, попробовать себя в'
+        ' дизайне и творчестве и узнать больше об осознанном потреблении.'
+    ),
+    'msg_4': (
+        'В 2020 году Фонд запустил исследовательский проект по переработке'
+        ' синтетической ткани в агломерат (сырьё, похожее на пластиковые'
+        ' гранулы, из которых можно произвести новые материалы, например'
+        ' ручки или пластиковые вешалки). Производство осуществляется через'
+        ' ООО «РЕКЛО» — дочку Фонда, которая развивает этот проект.  \n\n'
+        '*О реализации этих проектов и не только, мы рассказываем в годовых'
+        ' отчетах Фонда. Прислать тебе ссылку?*'
+    ),
+}
+
+ANNUAL_REPORTS = (
+    f'*Почитай наши [годовые отчеты]({ANNUAL_REPORTS_LINK})\.'
+    ' Мы ими очень гордимся\!*'
+)
