+++ src/bot/constants/state.py
@@ -1,5 +1,5 @@
 CHECK: int = 1
-MENU: int = 2
+MAIN_MENU: int = 2
 ABOUT_FUND_MENU_STATE: int = 60
 FUND_MISSION_STATE: int = 601
 THINGS_PATH_STATE: int = 602
